Remove duplicate path detail deserialization, fixes #1137
